configure liqpay
configure liqpay, telegramId, price type etc

+++ src-backend/YaHoApiService/YaHoApiService.BLL.Contracts/DTO/ViewData/Product/ProductViewData.cs
@@ -10,9 +10,9 @@ namespace YaHo.YaHoApiService.BLL.Contracts.DTO.ViewData.Product
 
         public int OrderId { get; set; }
 
-        public int Price { get; set; }
+        public decimal Price { get; set; }
 
-        public int Tax { get; set; }
+        public decimal Tax { get; set; }
 
         public string Description { get; set; }
