+++ src-backend/YaHoApiService/YaHoApiService/ViewModels/UserViewModels/Get/GetUserInfoViewModel.cs
@@ -2,9 +2,9 @@
 {
     public class GetUserInfoViewModel : GetPartialUserInfoViewModel
     {
-        public int Balance { get; set; }
+        public decimal Balance { get; set; }
 
-        public int Hold { get; set; }
+        public decimal Hold { get; set; }
 
     }
 }
